+++ src/board_util.h
@@ -3,11 +3,11 @@
 
 #include "coincoin.h"
 
-board_msg_info *board_find_id(const Board *board, int id);
+board_msg_info *board_find_id(const Board *board, int64_t id);
 board_msg_info *board_find_md5(const Board *board, md5_byte_t md5[16]);
 board_msg_info *boards_find_id(const Boards *boards, id_type id);
 id_type boards_last_id(const Boards *b);
-board_msg_info* board_find_previous_from_id(const Board *board, int id);
+board_msg_info* board_find_previous_from_id(const Board *board, int64_t id);
 board_msg_info *board_find_previous(const Board *board, board_msg_info *mi);
 char *board_get_tok(const unsigned char **p, const unsigned char **np, 
 		    unsigned char *tok, int max_toklen, int *has_initial_space);
